// src/store/settings.enum.ts
export enum Setting {
	CB_USERNAME = 'cbUsername',
	CB_PASSWORD = 'CB_PASSWORD',

	PROJECT_ID = 'projectId',
	CB_ADDRESS = 'cbAddress',
	INBOX_TRACKER_ID = 'inboxTrackerId',

	ADVANCED_SEARCH_ENABLED = 'ADVANCED_SEARCH_ENABLED',
	SELECTED_TRACKER = 'SELECTED_TRACKER',
	CBQL_STRING = 'CBQL_STRING',
	SUBQUERY_LINK_METHOD = 'SUBQUERY_LINK_METHOD',
	FILTER_CRITERIA = 'FILTER_CRITERIA',

	USER_MAPPING = 'USER_MAPPING',
	IMPORT_CONFIGURATION = 'importConfiguration',
}